refactor the script to be frame number rather then  frame range
